Major refactor of vulkan loader generator
Changing the way extensions are loaded to account for extensions
appearing in multiple places. This should improve robustness when the
vulkan header version does not match the exact version of the spec used
to generate the loader.

Added unit tests.

Now using cmake-format and clang-format to keep things pretty
(why is clang-format so limited yet so popular? there are too
many clang-format off directives.)